requirements.txt
pandas==3.0.1
python-dotenv==1.2.1
SQLAlchemy==2.0.46
streamlit==1.54.0
